Move ToolbarItems to their own file.
