+++ niftyinitiativetracker/niftyinitiativetracker/ContentView.swift
@@ -111,40 +111,6 @@ struct AddNewCreature: View {
     }
 }
 
-struct ToolbarItems: ToolbarContent {
-    
-    @Binding var creatures: [Creature]
-    
-    @Binding var editMode: Bool
-    
-    @Binding var isAddingNewCreature: Bool
-    
-    var body: some ToolbarContent {
-        ToolbarItemGroup {
-            HStack {
-                Button {
-                    editMode.toggle()
-                } label: {
-                    Text("Edit")
-                }.disabled(creatures.isEmpty)
-                Spacer()
-                Button {
-                    creatures.sort {
-                        ($0.initiativeRoll + $0.modifier) > ($1.initiativeRoll + $1.modifier)
-                    }
-                } label: {
-                    Image(systemName: "line.3.horizontal.decrease")
-                }.disabled(creatures.isEmpty)
-                Button {
-                    isAddingNewCreature = true
-                } label: {
-                    Image(systemName: "plus")
-                }
-            }
-        }
-    }
-}
-
 struct ContentView_Previews: PreviewProvider {
     static var previews: some View {
         ContentView()
